requirements.txt
# Production requirements (Linux compatible - Fly.io deployment)
# Note: MLX removed (macOS-only), uses NumPy inference in production

# Core ML
numpy>=1.24.0
safetensors>=0.4.0

# Database
asyncpg>=0.29.0

# Web/API
aiohttp>=3.9.0
websockets>=12.0
requests>=2.31.0
python-socks[asyncio]>=2.4.0  # SOCKS5 proxy for residential IP routing

# Dashboard
flask>=3.0.0
flask-socketio>=5.3.0
python-socketio>=5.10.0
eventlet>=0.35.0
gevent>=24.2.1
gevent-websocket>=0.10.1

# Utils
python-dotenv>=1.0.0

# Polymarket Trading
py-clob-client>=0.29.0

# Blockchain interaction (for profit transfers)
web3>=6.0.0
eth-account>=0.10.0
